+++ src/deleteWorkspace.ts
@@ -1,12 +1,17 @@
-const vscode = require("vscode");
+import vscode from "vscode";
+import WorkspaceTreeDataProvider from "./workspaceTreeDataProvider";
+import WorkspaceTreeItem from "./workspaceTreeItem";
 
 // Prompts the user for confirmation and then deletes a workspace file.
-const deleteWorkspace = async (context, treeDataProvider) => {
+export default async function (
+  context: WorkspaceTreeItem,
+  treeDataProvider: WorkspaceTreeDataProvider,
+) {
   const workspaceUri = vscode.Uri.file(context.workspaceFileNameAndFilePath);
 
   const results = await vscode.window.showWarningMessage(
     `Are you sure you want to delete ${context.label} workspace?`,
-    ...["Delete Workspace", "Cancel"]
+    ...["Delete Workspace", "Cancel"],
   );
 
   if (results === undefined || results === "Cancel") {
@@ -18,6 +23,4 @@ const deleteWorkspace = async (context, treeDataProvider) => {
 
   // Refresh Tree
   treeDataProvider.refresh();
-};
-
-module.exports = deleteWorkspace;
+}