+++ src/resolveTemplatePath.ts
@@ -1,26 +1,24 @@
-const process = require("process");
-
-const path = require("path");
-
-const fs = require("fs");
+import process from "node:process";
+import path from "node:path";
+import fs from "node:fs";
 
 const templatePattern = /(?<template>\${(?<type>env):(?<variable>[a-zA-Z_]+)})/;
 
 /* eslint max-classes-per-file: ["error", 2] */
 
-class InvalidTemplateStringError extends Error {
+export class InvalidTemplateStringError extends Error {
   constructor() {
     super("Invalid Template String");
   }
 }
 
 class FolderNotFoundError extends Error {
-  constructor(message) {
+  constructor(message: string) {
     super(`Folder was not Found: ${message}`);
   }
 }
 
-const resolveEnvVar = (variable) => {
+const resolveEnvVar = (variable: string) => {
   const value = process.env[variable];
   if (value === undefined) {
     throw new InvalidTemplateStringError();
@@ -28,7 +26,12 @@ const resolveEnvVar = (variable) => {
   return value;
 };
 
-const replaceMatch = (match, group1, group2, group3) => {
+const replaceMatch = (
+  match: string,
+  group1: string,
+  group2: string,
+  group3: string,
+) => {
   if (group2 === "env") {
     return resolveEnvVar(group3);
   }
@@ -40,13 +43,13 @@ const replaceMatch = (match, group1, group2, group3) => {
  * @param {string} filepath - unresolved path
  * @returns {string}
  */
-const resolveTemplatePath = (filepath) => {
+export default function (filepath: string): string {
   let templateString = filepath;
   while (true) {
     /* eslint no-constant-condition: "off" */
     const resolvedString = templateString.replace(
       templatePattern,
-      replaceMatch
+      replaceMatch,
     );
     if (templateString === resolvedString) {
       const resolvedPath = path.resolve(resolvedString);
@@ -57,6 +60,4 @@ const resolveTemplatePath = (filepath) => {
     }
     templateString = resolvedString;
   }
-};
-
-module.exports = resolveTemplatePath;
+}